update menu choices

+++ src/product/Product.java
@@ -49,7 +49,7 @@ public class Product {
 	}
 
 
-	public void setCode(float code) {
+	public void setCode(int code) {
 		this.code = code;
 	}
